Separate auth script TS source files from executables
