+++ libs/auth/scripts/cac/cac-get-cert
@@ -1,4 +1,4 @@
 #!/usr/bin/env node
 
 require('esbuild-runner/register');
-require('./cac_get_cert').main(process.argv.slice(2));
+require('./src/cac_get_cert').main(process.argv.slice(2));
