+++ libs/auth/scripts/cac/configure-simulated-cac-card
@@ -1,4 +1,4 @@
 #!/usr/bin/env node
 
 require('esbuild-runner/register');
-require('./configure_simulated_cac_card').main();
+require('./src/configure_simulated_cac_card').main();
