+++ libs/auth/scripts/configure-java-card
@@ -1,4 +1,4 @@
 #!/usr/bin/env node
 
 require('esbuild-runner/register');
-require('./configure_java_card').main();
+require('./src/configure_java_card').main();
