+++ libs/auth/scripts/generate-dev-keys-and-certs
@@ -1,4 +1,4 @@
 #!/usr/bin/env node
 
 require('esbuild-runner/register');
-require('./generate_dev_keys_and_certs').main(process.argv.slice(2));
+require('./src/generate_dev_keys_and_certs').main(process.argv.slice(2));
