+++ libs/auth/scripts/program-system-administrator-java-card
@@ -1,4 +1,4 @@
 #!/usr/bin/env node
 
 require('esbuild-runner/register');
-require('./program_system_administrator_java_card').main();
+require('./src/program_system_administrator_java_card').main();
